Suggested Communities column now able to scroll
