+++ packages/commonwealth/client/styles/pages/user_dashboard/dashboard_communities_preview.scss
@@ -43,6 +43,10 @@
   float: right;
   top: 70px;
   z-index: 1;
+  flex-grow: 1;
+  overflow-y: auto;
+  max-height: 100vh;
+  padding-bottom: 100px;
 
   @include mediumSmallInclusive {
     display: none;
